# 17-gin-swagger/routes.go
package main

import (
	"github.com/gin-gonic/gin"
	"webapp/handlers"
)

func CORSMiddleware() gin.HandlerFunc {
	return func(c *gin.Context) {
		c.Writer.Header().Set("Access-Control-Allow-Origin", "*")
		c.Writer.Header().Set("Access-Control-Allow-Credentials", "true")
		c.Writer.Header().Set("Access-Control-Allow-Headers", "Content-Type, Content-Length, Accept-Encoding, X-CSRF-Token, Authorization, accept, origin, Cache-Control, X-Requested-With")
		c.Writer.Header().Set("Access-Control-Allow-Methods", "POST, OPTIONS, GET, PUT")

		if c.Request.Method == "OPTIONS" {
			c.AbortWithStatus(204)
			return
		}

		c.Next()
	}
}

func initRoute() {
	router.Use(CORSMiddleware())

	personRoutes := router.Group("/persons")
	{
		personRoutes.GET("/:name", handlers.GetPerson)

		personRoutes.POST("", handlers.AddPerson)
	}
}